data-transforms: add jq template
This adds a transform for JQ so you can write a single expression and
deploy it to work on transformed values. There are plenty of other knobs
and bits that we could expose here, but this is a good start.

There is a seperate process that compiles `libjq` into a WebAssembly
library. This can be found at https://github.com/rockwotj/jq

There is a quick C++23 transform SDK that I whipped up, mostly to play
with C++23. Note that WebAssembly currently does not support C++
exceptions, so the code is written with that being disabled in mind. The
API for the C++ sdk could be better, this is by no means an "offcial"
SDK.

Then there is a small program that hooks up the jq library and the SDK
to create a pluggable filter.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>